Added some functionality, fixed some bugs, one or two issues outstanding.
new:

* load multiple bitmaps from world file - use 'bitmap' token to generate a
fixedobstacle.  'environment' is deprecated. bitmap entities obey color &
pose settings. Added worlds/multi-environment.world as an example. Check
out the view from the visiondevice! a fixedobstacle with default settings
borders the matrix but is not drawn in the GUI. bitmaps can be hidden like
other objects from the view-menu.

* the 'scale' keyword is replaced by 'resolution <meters per pixel>'.
In section 0 (top level) of the world file, it specifies the size in
meters of a matrix cell. In a 'bitmap' section, it specifies the size in
meters of a bitmap pixel. (using 'scale' generates a warning)

* view-menu option to visualize matrix occupancy in the GUI - great for
debugging, but also useful to check if your resolution is sensible.

* removed 4cm center of rotation offset from positiondevice. added this
definition in worlds/pioneer.inc:

  define pioneer position
  (
    offset [-0.04 0.0]
  )

* keyword 'size [X Y]' at top level specifies the minimum world size in
meters. the world is automatically enlarged to fit any posed entities,
including bitmaps, so if you don't specify a 'size' it just fits neatly
around whatever is there.

* new model for adding devices based on a new class called 'Library'.
Instead of adding devices to the entityfactory, you define a static named
constructor and create a global instance of your object. A constructor
takes care of the registration. entityfactory is gone.

bugfix:

* rtk figures did not obey entity offsets - now they do bu see issue
below.

<forgot some...?>

issues:

* devices don't inherit their parent's player port number. have to think
about this one. it's a useful thing but breaks context-free grammar and
will require another hack to keep it in.

* matrix un-mapping doesn't work properly for objects with origin offsets.
I'll fix this ASAP.